requirements.txt
# Core ML Framework
torch>=2.0.0
torchvision>=0.15.0

# Streamlit Web App
streamlit>=1.28.0

# Image Processing
Pillow>=9.5.0
opencv-python>=4.8.0

# Data Science
numpy>=1.24.0
scikit-learn>=1.3.0

# Plotting (optional, for training)
matplotlib>=3.7.0

# Progress bars and utilities
tqdm>=4.65.0

# Collections (if using Counter)
collections-extended>=2.0.0